Improve support for additional files referenced by tests
Using the magic string `PathToAtomicsFolder` in a test when referencing
additional files that are needed by the test will result in the path
being modified to the actual location of the file referenced.

This currently only works for bundled and local tests. Support for
pulling down additional files referenced in tests from GitHub (when not
running in bundled mode) will be added in the future.